chore(storybook): update storybook and chromatic to latest version
