+++ package.json
@@ -65,7 +65,7 @@
     "@babel/core": "^7.23.2",
     "@babel/plugin-transform-runtime": "^7.23.2",
     "@babel/preset-env": "^7.23.2",
-    "@chromatic-com/storybook": "^4.0.0-next.16",
+    "@chromatic-com/storybook": "^5.2.1",
     "@eslint/js": "^10.0.1",
     "@originjs/vite-plugin-require-context": "^1.0.9",
     "@rollup/plugin-replace": "^6.0.2",
@@ -84,7 +84,7 @@
     "babel-jest": "^30.4.1",
     "babel-plugin-macros": "^2.4.2",
     "bootstrap": "^5.3.2",
-    "chromatic": "^6.17.3",
+    "chromatic": "^17.2.0",
     "commitizen": "^4.3.0",
     "cz-conventional-changelog": "^3.3.0",
     "eslint": "^10.4.0",
@@ -113,7 +113,8 @@
     "typescript": "5.9.3",
     "typescript-eslint": "^8.60.0",
     "vite": "7.3.5",
-    "vite-plugin-dts": "^5.0.2"
+    "vite-plugin-dts": "^5.0.2",
+    "vue-tsc": "^3.3.3"
   },
   "dependencies": {
     "@floating-ui/dom": "^1.7.6",
