+++ stories/Introduction.mdx
@@ -1,4 +1,4 @@
-import { Meta } from '@storybook/addon-docs';
+import { Meta } from "@storybook/addon-docs/blocks";
 
 <Meta title="Introduction" />
